tweak fix warning
../plugins/fd/fd_common.h:528:32: warning: format ‘%lld’ expects argument of type ‘long long int’, but argument 3 has type ‘int64_t’ {aka ‘long int’} [-Wformat=]
  528 |       l = fprintf(fp, "time=%lld level=%c key=%s name=%s vollen=%d vol=%s\n",

../plugins/fd/fd_common.h:528:66: warning: format ‘%d’ expects argument of type ‘int’, but argument 7 has type ‘size_t’ {aka ‘long unsigned int’} [-Wformat=]
  528 |       l = fprintf(fp, "time=%lld level=%c key=%s name=%s vollen=%d vol=%s\n",

+++ bacula/src/plugins/fd/fd_common.h
@@ -528,11 +528,11 @@ bool joblist::store_job(char *data)
 
    if (level == 'F') {
       l = fprintf(fp, "time=%lld level=%c key=%s name=%s vollen=%d vol=%s\n", 
-               now, level, key, name, strlen(data), data);
+               (long long)now, level, key, name, (int)strlen(data), data);
 
    } else {
       l = fprintf(fp, "time=%lld level=%c key=%s name=%s root=%s prev=%s vollen=%d vol=%s\n",
-               now, level, key, name, root, prev, strlen(data), data);
+            (long long)now, level, key, name, root, prev, (int)strlen(data), data);
    }
 
    unbash_spaces(data);
